requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
streamlit>=1.25.0
plotly>=5.15.0
Pillow>=8.3.0 